slider

+++ Project_TMP/tpm_pcm.h
@@ -10,7 +10,7 @@ extern uint8_t signal[];
 extern uint8_t speed;
 extern int32_t timer;
 extern uint32_t TIME;
-extern uint8_t play_index;
+extern uint16_t play_index;
 extern uint8_t playFlag;
 extern const uint16_t SAMPLES;
